Skeleton for day 24
Add skeleton unsolved solution for day 24 of 2023.

+++ src/AdventOfCode/Puzzles/Y2023/Day22.cs
@@ -75,7 +75,7 @@ public sealed class Day22 : Puzzle
 
         static Vector3 ParseVector(ReadOnlySpan<char> value)
         {
-            (float x, float y, float z) = value.AsNumberTriple<int>();
+            (float x, float y, float z) = value.AsNumberTriple<float>();
             return new(x, y, z);
         }
